Rework pending chances
checkHit automatically committing is problematic as we must not
commit misses in scenarios where we don't commit hits as this
results in the Actions struct having a different shape, breaking
transitions

checkHit now returns whether it thinks it should be saved, but
further computation (which should only happen in the event chance is
enabled) is required to know whether to commit or not. This
additional computation is not expected to have large performance
implications when chance is enabled because it consists of simple
boolean ops, and overall being able to consolidate numerous
redundant chance branches because of pending should result in a net
boon to performance for those computing transitions.

Replacing commit/clearPending -> save/commit is mostly window
dressing and will likely get reverted