set duration fields on Actions
we can't easily conclusively determine that thrash has ended
without additional information. as the user, if thrashing = .hidden
then can check `choices` to see if is forced (one move that is index
0 or 1 depending on -Dshowdown) to determine if ended. if the
opponent, when .hidden the result is either unknowable after 3 turns
or known to be .ended if it has been 4 turns

we must set duration fields to be .continuing the moment they begin
in order to sure transitions can be generated. technically the
duration roll is sufficient for uniqueness of the key/fully
describing the RNG, but to easily enable transitions to determine
which field to perturb we need to set that as well. another state in
the enum is unnecessary because .continuing + the durations field
being set is understood to convery .started