Revert "clear duration roll if the duration they were for gets cleared"
so we can't actually do this because while its functionally OK, the
battle bytes from an overridden vs. non-overridden battle won't
match because Actions will be underspecified. in essence, even
though the duration bytes don't mean anything given the volatiles
they correspond to have been cleared, bc the game is lazy in
clearing them we need to leave them around.

updating the code to clear them *shouldnt* effect correctness (but
might), but will impact performance for cases which dont care about
chance/calc/transitions, so we can't change things. If this starts
to matter in transitions (due to double counting states) we simply
need to make the roll helper smarter, checking to see a duration
field has been .started (which can be done with a single bit mask to
check for started | ended on the corresponding fields all at once
followed by one to filter out those that have the ended bit set).

This reverts commit 73ae24d52f118e60aab3aa1aeafa25ba4cfbad85.