Support all the way back to Zig v0.11.0
Required as a workaround for ziglang/zig#17768 - currently the
performance is around 20% slower on master which is unacceptable.
This is obviously somewhat of a bummer (RIP 3a62fd35) and will be
a pretty big maintenance headache but its been almost a year without
(positive) progress on the Zig side so there's not really any better
options.

Clean `make integration` on the following builds:

  0.11.0
  0.12.0
  0.12.0-dev.789+e6590fea1
  0.12.0-dev.866+3a47bc715
  0.12.0-dev.876+aaf46187a
  0.12.0-dev.1396+f6de3ec96
  0.12.0-dev.1879+e19219fa0
  0.12.0-dev.2036+fc79b22a9
  0.12.0-dev.2665+919a3bae1
  0.12.0-dev.3644+05d975576
  0.12.1
  0.13.0
  0.13.0-dev.8+c352845e8
  0.13.0-dev.39+f6f7a47aa
  0.13.0-dev.46+3648d7df1
  0.14.0-dev.23+d9bd34fd0
  0.14.0-dev.564+75cf7fca9
  0.14.0-dev.994+9f46abf59