ensure duration ranges are correct
in scenarios where theres a speed tie, one player used two different
duration moves on subsequent turns and the other uses Mirror Move
(and thus could proc either of the two duration moves), the
overridden duration could have the wrong range when underspecified.

removing the asserts might also "just work" (and would be faster),
but dealing with completely incorrect durations seems rather sketchy
(eg. what if we pass a sleep duration of 1 to the binding roll which
subtracts 1 from it and then makes it duration 0?)